employees/requirements.txt
appdirs==1.4.3
asgiref==3.2.5
attrs==19.3.0
boto3==1.12.39
botocore==1.15.39
Brotli==1.0.7
certifi==2019.11.28
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.4
distlib==0.3.0
Django==3.0.7
django-compression-middleware==0.4.1
django-cors-headers==3.2.1
django-extensions==2.2.9
django-filter==2.2.0
django-import-csv==0.1.10
django-rest-swagger==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
drf-yasg==1.17.1
et-xmlfile==1.0.1
filelock==3.0.12
fsspec==0.7.2
gunicorn==20.0.4
identify==1.4.13
idna==2.9
inflection==0.3.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
more-itertools==8.2.0
nodeenv==1.3.5
nose==1.3.7
numpy==1.18.2
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==20.3
pandas==1.0.3
Pillow==7.1.2
pluggy==0.13.1
pre-commit==2.2.0
psycopg2-binary==2.8.4
py==1.8.1
pycparser==2.20
pyparsing==2.4.6
pytest==5.4.1
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3fs==0.4.2
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.13
wcwidth==0.1.9
zstandard==0.13.0
